+++ sdk/Managed/src/Microsoft.WindowsAzure.MobileServices.Android/Authentication/MobileServiceUIAuthentication.cs
@@ -2,7 +2,7 @@ using System;
 using System.Globalization;
 using Android.Content;
 using System.Threading.Tasks;
-using Xamarin.Auth;
+using Xamarin.Auth._MobileServices;
 using System.Collections.Generic;
 
 namespace Microsoft.WindowsAzure.MobileServices